Make alpha a int

+++ data_struct/rgba_rect.py
@@ -1,7 +1,7 @@
 
 class RGBARect:
 
-    def __init__(self, r: int, g: int, b: int, a: float, x: int, y: int, w: int, h: int):
+    def __init__(self, r: int, g: int, b: int, a: int, x: int, y: int, w: int, h: int):
         self.r = r
         self.g = g
         self.b = b
